+++ Changelog.md
@@ -1,5 +1,11 @@
 ## tag: V6 (WIP)
 
+### 1. Get statement support.
+
+* GetOp: Grammar, builder, runtime, and prototype generator helper functions for lowering, and lowering implementation.
+* printOp lowering: Merge bulk of createToyPrintF64Prototype, createToyPrintI64Prototype functions into a single generic helper function.
+
+
 ## tag: V5 (Dec 22, 2025)
 
 The language now supports functions, calls, parameters, returns, and basic conditional blocks.